# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{projectId}/contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List contracts for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"
    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a contract
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /projects/{projectId}/contracts/{id}:
    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update a contract
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractBody"
      responses:
        "200":
          description: Contract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects/{projectId}/photos:
    get:
      operationId: listPhotos
      tags: [photos]
      summary: List photos for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: createPhoto
      tags: [photos]
      summary: Upload a photo (base64)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhotoBody"
      responses:
        "201":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /projects/{projectId}/photos/{id}:
    patch:
      operationId: updatePhoto
      tags: [photos]
      summary: Update a photo
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhotoBody"
      responses:
        "200":
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
    delete:
      operationId: deletePhoto
      tags: [photos]
      summary: Delete a photo
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Project:
      type: object
      properties:
        id:
          type: integer
        builderId:
          type: integer
        clientId:
          type: ["integer", "null"]
        clientName:
          type: string
        clientEmail:
          type: ["string", "null"]
        name:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        startDate:
          type: ["string", "null"]
          format: date
        notes:
          type: ["string", "null"]
        progress:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - builderId
        - clientId
        - clientName
        - clientEmail
        - name
        - address
        - status
        - startDate
        - notes
        - progress
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        clientName:
          type: string
        clientEmail:
          type: ["string", "null"]
        name:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [planning, active, on_hold, completed, cancelled]
        startDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        progress:
          type: integer
      required:
        - clientName
        - name
        - address
        - status
    UpdateProjectBody:
      type: object
      properties:
        clientName:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [planning, active, on_hold, completed, cancelled, "null"]
        startDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        progress:
          type: ["integer", "null"]
    Contract:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        fileUrl:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, signed]
        uploadedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - title
        - fileUrl
        - version
        - status
        - uploadedAt
    CreateContractBody:
      type: object
      properties:
        title:
          type: string
        fileUrl:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, sent, signed]
      required:
        - title
        - status
    UpdateContractBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        fileUrl:
          type: ["string", "null"]
        version:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
          enum: [draft, sent, signed, "null"]
...
    Photo:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        fileUrl:
          type: string
        caption:
          type: ["string", "null"]
        visibleToClient:
          type: boolean
        uploadedBy:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - fileUrl
        - caption
        - visibleToClient
        - uploadedBy
        - createdAt
    CreatePhotoBody:
      type: object
      properties:
        fileUrl:
          type: string
        caption:
          type: ["string", "null"]
        visibleToClient:
          type: boolean
      required:
        - fileUrl
    UpdatePhotoBody:
      type: object
      properties:
        caption:
          type: ["string", "null"]
        visibleToClient:
          type: ["boolean", "null"]
...