# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContractBodyStatus } from "./createContractBodyStatus";

export interface CreateContractBody {
  title: string;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  version?: string | null;
  status: CreateContractBodyStatus;
}