# lib/api-zod/src/generated/types/createPhotoBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePhotoBody {
  fileUrl: string;
  /** @nullable */
  caption?: string | null;
  visibleToClient?: boolean;
}