# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BuildOS API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProjectBodyStatus } from "./updateProjectBodyStatus";

export interface UpdateProjectBody {
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  status?: UpdateProjectBodyStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  progress?: number | null;
}